requirements.txt
# NYC Tax RAG System Dependencies
# Python 3.10+ required

# LangChain ecosystem
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5

# Vector database
chromadb>=0.4.0

# Keyword search
rank_bm25>=0.2.2

# Environment management
python-dotenv>=1.0.0

# Web interface
streamlit>=1.30.0

# Tokenization
tiktoken>=0.5.0

# Data validation
pydantic>=2.0.0

# Progress bar
tqdm>=4.66.0
